Worked on hold mouse button so that it moves towards the mouse.

+++ src/drawcircle/CircleObject.java
@@ -7,32 +7,32 @@ import java.awt.Color;
 public class CircleObject {
 
     
-    private int x;
-    private int y;
+    private double x;
+    private double y;
     private int radius;
     private Color color;
     public CircleObject () {
         
     }
-    public CircleObject(int x, int y, int radius, Color color) {
+    public CircleObject(double x, double y, int radius, Color color) {
         this.x = x;
         this.y = y;
         this.radius = radius;
         this.color = color;
     }
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
